backend/requirements.txt
fastapi==0.128.0
uvicorn==0.35.0
sqlalchemy==2.0.38
psycopg[binary]==3.2.3
pydantic==2.12.0
pydantic-settings==2.10.1
python-dotenv==1.1.1
requests==2.32.4
beautifulsoup4==4.13.4
langchain==0.1.17
langchain-google-genai==0.0.6
google-generativeai==0.3.2
lxml
gunicorn
uvicorn[standard]